+++ apps/ethereum_jsonrpc/lib/ethereum_jsonrpc.ex
@@ -9,12 +9,12 @@ defmodule EthereumJSONRPC do
       config :ethereum_jsonrpc,
         url: "http://localhost:8545",
         trace_url: "http://localhost:8545",
-        http: [recv_timeout: 60_000, timeout: 60_000, hackney: [pool: :ethereum_jsonrpc]]
+        http: [recv_timeout: 60_000, timeout: 60_000, pool: :ethereum_jsonrpc]
 
 
   Note: the tracing node URL is provided separately from `:url`, via `:trace_url`. The trace URL and is used for
-  `fetch_internal_transactions`, which is only a supported method on tracing nodes. The `:http` option is passed
-  directly to the HTTP library (`HTTPoison`), which forwards the options down to `:hackney`.
+  `fetch_internal_transactions`, which is only a supported method on tracing nodes. The `:http` option is adapted
+  to the HTTP library (`HTTPoison` or `Tesla.Mint`).
 
   ## Throttling