+++ apps/explorer/test/explorer/market/source/coin_market_cap_test.exs
@@ -28,9 +28,12 @@ defmodule Explorer.Market.Source.CoinMarketCapTest do
       currency_id: "2781"
     )
 
+    Application.put_env(:tesla, :adapter, Tesla.Adapter.Mint)
+
     on_exit(fn ->
       Application.put_env(:explorer, Explorer.Market.Source, source_configuration)
       Application.put_env(:explorer, CoinMarketCap, coin_market_cap_configuration)
+      Application.put_env(:tesla, :adapter, Explorer.Mock.TeslaAdapter)
     end)
 
     {:ok, bypass: bypass}