+++ apps/indexer/lib/indexer/fetcher/filecoin/beryx_api.ex
@@ -4,8 +4,7 @@ defmodule Indexer.Fetcher.Filecoin.BeryxAPI do
   address hash
   """
 
-  alias Explorer.Helper
-  alias HTTPoison.Response
+  alias Explorer.{Helper, HttpClient}
 
   @doc """
   Fetches account information for a given Ethereum address hash from the Beryx API.
@@ -16,12 +15,12 @@ defmodule Indexer.Fetcher.Filecoin.BeryxAPI do
   ## Returns
   - `{:ok, map()}`: On success, returns the account information as a map.
   - `{:error, integer(), map()}`: On failure, returns the HTTP status code and the error message as a map.
-  - `{:error, HTTPoison.Error.t()}`: On network or other HTTP errors, returns the error structure.
+  - `{:error, any()}`: On network or other HTTP errors, returns the error reason.
   """
   @spec fetch_address_info(EthereumJSONRPC.address()) ::
           {:ok, map()}
           | {:error, integer(), map()}
-          | {:error, HTTPoison.Error.t()}
+          | {:error, any()}
   def fetch_address_info(eth_address_hash) do
     config = Application.get_env(:indexer, __MODULE__)
     base_url = config |> Keyword.get(:base_url) |> String.trim_trailing("/")
@@ -34,16 +33,16 @@ defmodule Indexer.Fetcher.Filecoin.BeryxAPI do
       {"Content-Type", "application/json"}
     ]
 
-    case HTTPoison.get(url, headers) do
-      {:ok, %Response{body: body, status_code: 200}} ->
+    case HttpClient.get(url, headers) do
+      {:ok, %{body: body, status_code: 200}} ->
         json = Helper.decode_json(body)
         {:ok, json}
 
-      {:ok, %Response{body: body, status_code: status_code}} ->
+      {:ok, %{body: body, status_code: status_code}} ->
         json = Helper.decode_json(body)
         {:error, status_code, json}
 
-      {:error, %HTTPoison.Error{}} = error ->
+      error ->
         error
     end
   end